Had some fun.
Decided to have some fun with 8 and clean it up a bit.

In both, instead of the if statement that increments the pointer and
jumps back, I just told it to use a default sub of $ptr++ if one wasn't
provided.

In 8b, instead of creating a copy of the list, we localise the single
element in the array, mess with it, and then it will get automatically
reset by perl at the end of the lexical.